requirements.txt
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
bcrypt==3.1.5
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.4.2
future==0.17.1
idna==2.8
Jinja2==2.10
junos-eznc==2.2.0
lxml==4.3.0
MarkupSafe==1.1.0
more-itertools==5.0.0
mypy-extensions==0.4.1
napalm==2.3.3
ncclient==0.6.3
netaddr==0.7.19
netmiko==2.3.0
nornir==2.0.0
paramiko==2.4.2
pluggy==0.8.0
py==1.7.0
pyasn1==0.4.5
pycparser==2.19
pydantic==0.17
pyeapi==0.8.2
pyIOSXR==0.53
PyNaCl==1.3.0
pynxos==0.0.3
pyserial==3.4
pytest==4.0.2
PyYAML==3.13
requests==2.21.0
ruamel.yaml==0.15.83
scp==0.13.0
selectors2==2.0.1
six==1.12.0
textfsm==0.4.1
urllib3==1.24.1
